Add memoization to cache the results of current_organisation
'current_organisation' method is getting called inside the loop which results in repeated  queries and slowing down the page load.

To optimise the log search/filters, 'current_organisation` method needs to be memoized. It is good candidate for  memoization as it is being used in different pages of the application.